v4

+++ src/main/java/com/example/libreria/repo/LibroRepo.java
@@ -8,10 +8,10 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface LibroRepo extends JpaRepository<Libro, Integer> {
-    void deleteLibroById(int id);
+public interface LibroRepo extends JpaRepository<Libro, Long> {
+    void deleteLibroById(Long id);
 
-    Optional<Libro> findLibroById(int id);
+    Optional<Libro> findLibroById(Long id);
     List<Libro> findLibroByAutorDni(String dni);
-    List<Libro> findLibroByCategoriaId(int id);
+    List<Libro> findLibroByCategoriaId(Long id);
 }
